// src/components/__generated__/TripReviews.types.ts
import * as GraphQLTypes from "../../__generated__/types";

export type Review_reviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type HostAndLocationReviewFragment_locationReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type HostAndLocationReviewFragment_hostReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type HostAndLocationReviewFragment_guestReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type GuestReview_locationReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type GuestReview_hostReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};

export type GuestReview_guestReviewFragment = {
  __typename: "Review";
  rating: number;
  text: string;
};